@@ -1,7 +1,6 @@
 <?php
-
 /**
- * ZeroOrMore.php
+ * CategoryStoreRequest.php
  * Copyright (c) 2019 james@firefly-iii.org
  *
  * This file is part of Firefly III (https://github.com/firefly-iii).
@@ -22,44 +21,44 @@
 
 declare(strict_types=1);
 
-namespace FireflyIII\Rules;
+namespace FireflyIII\Api\V1\Requests;
 
-use Illuminate\Contracts\Validation\Rule;
+use FireflyIII\Rules\ZeroOrMore;
+use FireflyIII\Support\Request\ChecksLogin;
+use FireflyIII\Support\Request\ConvertsDataTypes;
+use Illuminate\Foundation\Http\FormRequest;
 
 /**
+ * Class CategoryRequest
  *
- * Class ZeroOrMore
  * @codeCoverageIgnore
  */
-class ZeroOrMore implements Rule
+class CategoryStoreRequest extends FormRequest
 {
+    use ConvertsDataTypes, ChecksLogin;
 
     /**
-     * Get the validation error message.
+     * Get all data from the request.
      *
-     * @return string
+     * @return array
      */
-    public function message(): string
+    public function getAll(): array
     {
-        return trans('validation.zero_or_more');
+        return [
+            'name'  => $this->string('name'),
+            'notes' => $this->nlString('notes'),
+        ];
     }
 
     /**
-     * Determine if the validation rule passes.
+     * The rules that the incoming request must be matched against.
      *
-     * @param  string $attribute
-     * @param  mixed  $value
-     *
-     * @return bool
+     * @return array
      */
-    public function passes($attribute, $value): bool
+    public function rules(): array
     {
-        $value = (string)$value;
-        if ('' === $value) {
-            return true;
-        }
-        $res = bccomp('0', $value);
-
-        return $res <= 0;
+        return [
+            'name' => 'required|between:1,100|uniqueObjectForUser:categories,name',
+        ];
     }
 }
