@@ -27,6 +27,7 @@ use DB;
 use FireflyIII\Models\Transaction;
 use FireflyIII\Models\TransactionJournal;
 use Illuminate\Console\Command;
+use Log;
 use stdClass;
 
 /**
@@ -55,6 +56,7 @@ class FixUnevenAmount extends Command
      */
     public function handle(): int
     {
+        Log::debug(sprintf('Now in %s', __METHOD__));
         $start = microtime(true);
         $count = 0;
         // get invalid journals
@@ -64,8 +66,11 @@ class FixUnevenAmount extends Command
                       ->get(['transaction_journal_id', DB::raw('SUM(amount) AS the_sum')]);
         /** @var stdClass $entry */
         foreach ($journals as $entry) {
-            if (0 !== bccomp((string) $entry->the_sum, '0')) {
-                $this->fixJournal((int) $entry->transaction_journal_id);
+            if (0 !== bccomp((string)$entry->the_sum, '0')) {
+                $message = sprintf('Sum of journal #%d is %s instead of zero.', $entry->transaction_journal_id, $entry->the_sum);
+                $this->warn($message);
+                Log::warning($message);
+                $this->fixJournal((int)$entry->transaction_journal_id);
                 $count++;
             }
         }
@@ -106,7 +111,7 @@ class FixUnevenAmount extends Command
             return;
         }
 
-        $amount = bcmul('-1', (string) $source->amount);
+        $amount = bcmul('-1', (string)$source->amount);
 
         // fix amount of destination:
         /** @var Transaction $destination */
