@@ -23,6 +23,7 @@ declare(strict_types=1);
 namespace FireflyIII\Http\Requests;
 
 use FireflyIII\Support\Request\ChecksLogin;
+use FireflyIII\Support\Request\ConvertsDataTypes;
 use Illuminate\Foundation\Http\FormRequest;
 
 /**
@@ -32,7 +33,7 @@ use Illuminate\Foundation\Http\FormRequest;
  */
 class EmailFormRequest extends FormRequest
 {
-    use ChecksLogin;
+    use ChecksLogin, ConvertsDataTypes;
 
     /**
      * Rules for this request.
