@@ -1,2 +1,2 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[2],{68:function(e,n,t){e.exports=t(73)},73:function(e,n,t){"use strict";t.r(n);var r=t(6),o={name:"Index"},u=t(0),c=Object(u.a)(o,(function(){var e=this.$createElement,n=this._self._c||e;return n("div",[n("router-view")],1)}),[],!1,null,"6b40cc75",null).exports;t(20);var s=[{path:"/",component:c}],i=new r.a({mode:"history",routes:s}),a=t(18),p={};Vue.use(r.a),new Vue({router:i,i18n:a,render:function(e){return e(c,{props:p})}}).$mount("#accounts")}},[[68,0,1]]]);
+(window.webpackJsonp=window.webpackJsonp||[]).push([[2],{221:function(e,n,t){e.exports=t(225)},225:function(e,n,t){"use strict";t.r(n);var r=t(23),o={name:"Index"},u=t(1),c=Object(u.a)(o,(function(){var e=this.$createElement,n=this._self._c||e;return n("div",[n("router-view")],1)}),[],!1,null,"6b40cc75",null).exports;t(10);var s=[{path:"/",component:c}],i=new r.a({mode:"history",routes:s}),a=t(167),p={};Vue.use(r.a),new Vue({router:i,i18n:a,render:function(e){return e(c,{props:p})}}).$mount("#accounts")}},[[221,0,1]]]);
 //# sourceMappingURL=index.js.map
