@@ -1,2 +1,2 @@
-(window.webpackJsonp=window.webpackJsonp||[]).push([[3],{21:function(t,a,s){s(72),t.exports=s(74)},72:function(t,a,s){"use strict";s.r(a);var e={name:"Dashboard"},n=s(0),i=Object(n.a)(e,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",[s("top-boxes"),t._v(" "),s("div",{staticClass:"row"},[s("div",{staticClass:"col-lg-12 col-md-12 col-sm-12 col-xs-12"},[s("main-account")],1)]),t._v(" "),s("main-account-list"),t._v(" "),s("div",{staticClass:"row"},[s("div",{staticClass:"col-lg-6 col-md-6 col-sm-12 col-xs-12"},[s("main-budget")],1),t._v(" "),s("div",{staticClass:"col-lg-6 col-md-6 col-sm-12 col-xs-12"},[s("main-category")],1)]),t._v(" "),s("div",{staticClass:"row"},[s("div",{staticClass:"col-lg-8 col-md-8 col-sm-12 col-xs-12"},[s("main-debit")],1),t._v(" "),s("div",{staticClass:"col-lg-4 col-md-4 col-sm-12 col-xs-12"},[s("main-credit")],1)]),t._v(" "),s("div",{staticClass:"row"},[s("div",{staticClass:"col-lg-6 col-md-6 col-sm-12 col-xs-12"},[s("main-piggy-list")],1),t._v(" "),s("div",{staticClass:"col-lg-6 col-md-6 col-sm-12 col-xs-12"},[s("main-bills-list")],1)])],1)}),[],!1,null,null,null).exports,r={name:"TopBoxes",data:function(){return{summary:[],balances:[],billsPaid:[],billsUnpaid:[],leftToSpend:[],netWorth:[]}},mounted:function(){this.prepareComponent()},methods:{prepareComponent:function(){var t=this;axios.get("./api/v1/summary/basic?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(a){t.summary=a.data,t.buildComponent()}))},buildComponent:function(){this.getBalanceEntries(),this.getBillsEntries(),this.getLeftToSpend(),this.getNetWorth()},getBalanceEntries:function(){this.balances=this.getKeyedEntries("balance-in-")},getNetWorth:function(){this.netWorth=this.getKeyedEntries("net-worth-in-")},getLeftToSpend:function(){this.leftToSpend=this.getKeyedEntries("left-to-spend-in-")},getBillsEntries:function(){this.billsPaid=this.getKeyedEntries("bills-paid-in-"),this.billsUnpaid=this.getKeyedEntries("bills-unpaid-in-")},getKeyedEntries:function(t){var a=[];for(var s in this.summary)this.summary.hasOwnProperty(s)&&t===s.substr(0,t.length)&&a.push(this.summary[s]);return a}}},c=Object(n.a)(r,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"row"},[s("div",{staticClass:"col-md-3 col-sm-6 col-12"},[s("div",{staticClass:"info-box"},[t._m(0),t._v(" "),s("div",{staticClass:"info-box-content"},[s("span",{staticClass:"info-box-text"},[t._v(t._s(t.$t("firefly.balance")))]),t._v(" "),t.balances.length>0?s("span",{staticClass:"info-box-number"},[t._v(t._s(t.balances[0].value_parsed))]):t._e(),t._v(" "),t._m(1),t._v(" "),s("span",{staticClass:"progress-description"},t._l(t.balances,(function(a){return s("span",[t._v(t._s(a.sub_title)),s("br")])})),0)])])]),t._v(" "),s("div",{staticClass:"col-12 col-sm-6 col-md-3"},[s("div",{staticClass:"info-box"},[t._m(2),t._v(" "),s("div",{staticClass:"info-box-content"},[s("span",{staticClass:"info-box-text"},[s("span",[t._v(t._s(t.$t("firefly.bills_to_pay")))])]),t._v(" "),1===t.billsUnpaid.length&&t.billsPaid.length>0?s("span",{staticClass:"info-box-number"},[t._v(t._s(t.billsUnpaid[0].value_parsed))]):t._e(),t._v(" "),t._m(3),t._v(" "),s("span",{staticClass:"progress-description"},[1===t.billsUnpaid.length&&1===t.billsPaid.length?s("span",[t._v(t._s(t.$t("firefly.paid"))+": "+t._s(t.billsPaid[0].value_parsed))]):t._e(),t._v(" "),t.billsUnpaid.length>1?s("span",t._l(t.billsUnpaid,(function(a,e){return s("span",{key:a.key},[t._v("\n                            "+t._s(a.value_parsed)),e+1!==t.billsUnpaid.length?s("span",[t._v(", ")]):t._e()])})),0):t._e()])])])]),t._v(" "),s("div",{staticClass:"clearfix hidden-md-up"}),t._v(" "),s("div",{staticClass:"col-12 col-sm-6 col-md-3"},[s("div",{staticClass:"info-box"},[t._m(4),t._v(" "),s("div",{staticClass:"info-box-content"},[s("span",{staticClass:"info-box-text"},[s("span",[t._v(t._s(t.$t("firefly.left_to_spend")))])]),t._v(" "),t.leftToSpend.length>0?s("span",{staticClass:"info-box-number"},[t._v(t._s(t.leftToSpend[0].value_parsed))]):t._e(),t._v(" "),t._m(5),t._v(" "),s("span",{staticClass:"progress-description"},t._l(t.leftToSpend,(function(a,e){return s("span",{key:a.key},[t._v("\n                            "+t._s(a.value_parsed)),e+1!==t.leftToSpend.length?s("span",[t._v(", ")]):t._e()])})),0)])])]),t._v(" "),s("div",{staticClass:"col-12 col-sm-6 col-md-3"},[s("div",{staticClass:"info-box"},[t._m(6),t._v(" "),s("div",{staticClass:"info-box-content"},[s("span",{staticClass:"info-box-text"},[s("span",[t._v(t._s(t.$t("firefly.net_worth")))])]),t._v(" "),t.netWorth.length>0?s("span",{staticClass:"info-box-number"},[t._v(t._s(t.netWorth[0].value_parsed))]):t._e(),t._v(" "),t._m(7),t._v(" "),s("span",{staticClass:"progress-description"},t._l(t.netWorth,(function(a,e){return s("span",{key:a.key},[t._v("\n                            "+t._s(a.value_parsed)),e+1!==a.length?s("span",[t._v(", ")]):t._e()])})),0)])])])])}),[function(){var t=this.$createElement,a=this._self._c||t;return a("span",{staticClass:"info-box-icon"},[a("i",{staticClass:"far fa-bookmark text-info"})])},function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"progress bg-info"},[a("div",{staticClass:"progress-bar",staticStyle:{width:"0"}})])},function(){var t=this.$createElement,a=this._self._c||t;return a("span",{staticClass:"info-box-icon"},[a("i",{staticClass:"far fa-calendar-alt text-teal"})])},function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"progress bg-teal"},[a("div",{staticClass:"progress-bar",staticStyle:{width:"0"}})])},function(){var t=this.$createElement,a=this._self._c||t;return a("span",{staticClass:"info-box-icon"},[a("i",{staticClass:"fas fa-money-bill text-success"})])},function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"progress bg-success"},[a("div",{staticClass:"progress-bar",staticStyle:{width:"0"}})])},function(){var t=this.$createElement,a=this._self._c||t;return a("span",{staticClass:"info-box-icon"},[a("i",{staticClass:"fas fa-money-bill text-success"})])},function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"progress bg-success"},[a("div",{staticClass:"progress-bar",staticStyle:{width:"0"}})])}],!1,null,"2b5ab8b3",null).exports,o={name:"DataConverter",data:function(){return{dataSet:null,newDataSet:null}},methods:{convertChart:function(t){return this.dataSet=t,this.newDataSet={count:0,labels:[],datasets:[]},this.getLabels(),this.getDataSets(),this.newDataSet.count=this.newDataSet.datasets.length,this.newDataSet},colorizeBarData:function(t){this.dataSet=t,this.newDataSet={count:0,labels:[],datasets:[]};for(var a=[[53,124,165],[0,141,76],[219,139,11],[202,25,90],[85,82,153],[66,133,244],[219,68,55],[244,180,0],[15,157,88],[171,71,188],[0,172,193],[255,112,67],[158,157,36],[92,107,192],[240,98,146],[0,121,107],[194,24,91]],s=[],e=0;e<a.length;e++)s.push("rgba("+a[e][0]+", "+a[e][1]+", "+a[e][2]+", 0.5)");for(var n in this.newDataSet.labels=this.dataSet.labels,this.newDataSet.count=this.dataSet.count,this.dataSet.datasets)if(this.dataSet.datasets.hasOwnProperty(n)){var i=this.dataSet.datasets[n];i.fill=!1,i.backgroundColor=i.borderColor=s[n],this.newDataSet.datasets.push(i)}return this.newDataSet},colorizeLineData:function(t){this.dataSet=t,this.newDataSet={count:0,labels:[],datasets:[]};for(var a=[[53,124,165],[0,141,76],[219,139,11],[202,25,90],[85,82,153],[66,133,244],[219,68,55],[244,180,0],[15,157,88],[171,71,188],[0,172,193],[255,112,67],[158,157,36],[92,107,192],[240,98,146],[0,121,107],[194,24,91]],s=[],e=0;e<a.length;e++)s.push("rgba("+a[e][0]+", "+a[e][1]+", "+a[e][2]+", 0.5)");for(var n in this.newDataSet.labels=this.dataSet.labels,this.newDataSet.count=this.dataSet.count,this.dataSet.datasets)if(this.dataSet.datasets.hasOwnProperty(n)){var i=this.dataSet.datasets[n];i.fill=!1,i.backgroundColor=i.borderColor=s[n],this.newDataSet.datasets.push(i)}return this.newDataSet},convertLabelsToDate:function(t){for(var a in t.labels)if(t.labels.hasOwnProperty(a)){var s=Date.parse(t.labels[a]);t.labels[a]=new Intl.DateTimeFormat(window.localeValue).format(s)}return t},getLabels:function(){var t=this.dataSet[0];if(void 0!==t)for(var a in t.entries)t.entries.hasOwnProperty(a)&&this.newDataSet.labels.push(a)},getDataSets:function(){for(var t in this.dataSet)if(this.dataSet.hasOwnProperty(t)){var a={},s=this.dataSet[t];if(void 0!==s){for(var e in a.label=s.label,a.type=s.type,a.currency_symbol=s.currency_symbol,a.currency_code=s.currency_code,a.yAxisID=s.yAxisID,a.data=[],s.entries)s.entries.hasOwnProperty(e)&&a.data.push(s.entries[e]);this.newDataSet.datasets.push(a)}}}}},l=Object(n.a)(o,void 0,void 0,!1,null,null,null).exports,d={name:"DefaultLineOptions",data:function(){return{}},methods:{formatLabel:function(t,a){var s=[],e=(t=String(t)).split(" "),n="";return e.forEach((function(t,i){if(n.length>0){var r=n+" "+t;if(!(r.length>a))return i===e.length-1?void s.push(r):void(n=r);s.push(n),n=""}i!==e.length-1&&t.length<a?n=t:s.push(t)})),s},getDefaultOptions:function(){return{legend:{display:!1},animation:{duration:0},responsive:!0,maintainAspectRatio:!1,elements:{line:{cubicInterpolationMode:"monotone"}},scales:{xAxes:[{gridLines:{display:!1},ticks:{callback:function(t,a,s){return t}}}],yAxes:[{display:!0,ticks:{callback:function(t){var a=this.chart.data.datasets[0].currency_code?this.chart.data.datasets[0].currency_code:"EUR";return new Intl.NumberFormat(window.localeValue,{style:"currency",currency:a}).format(t)},beginAtZero:!0}}]},tooltips:{mode:"label",callbacks:{label:function(t,a){var s=a.datasets[t.datasetIndex].currency_code?a.datasets[t.datasetIndex].currency_code:"EUR",e=new Intl.NumberFormat(window.localeValue,{style:"currency",currency:s}).format(t.yLabel);return a.datasets[t.datasetIndex].label+": "+e}}}}}}},u=Object(n.a)(d,(function(){var t=this.$createElement;return(this._self._c||t)("div")}),[],!1,null,"11bcfc89",null).exports,_={name:"MainAccount",mounted:function(){axios.get("./api/v1/chart/account/overview?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(t){var a=l.methods.convertChart(t.data);a=l.methods.colorizeLineData(a);var s=$("#mainAccountsChart").get(0).getContext("2d");new Chart(s,{type:"line",data:a,options:u.methods.getDefaultOptions()})}))}},f=Object(n.a)(_,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.yourAccounts")))])]),t._v(" "),t._m(0),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./accounts/asset"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_asset_accounts")))])])])}),[function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"card-body"},[a("div",[a("canvas",{staticStyle:{"min-height":"400px",height:"400px","max-height":"400px","max-width":"100%"},attrs:{id:"mainAccountsChart"}})])])}],!1,null,null,null).exports,v={name:"MainAccountList",data:function(){return{accounts:[]}},mounted:function(){var t=this;axios.get("./api/v1/preferences/frontpageAccounts").then((function(a){t.loadAccounts(a)}))},methods:{loadAccounts:function(t){var a=t.data.data.attributes.data;for(var s in a)a.hasOwnProperty(s)&&/^0$|^[1-9]\d*$/.test(s)&&s<=4294967294&&(this.accounts.push({id:a[s],title:"",uri:"",transactions:[]}),this.loadSingleAccount(s,a[s]))},loadSingleAccount:function(t,a){var s=this;axios.get("./api/v1/accounts/"+a).then((function(e){s.accounts[t].title=e.data.data.attributes.name,s.accounts[t].uri="./accounts/show/"+e.data.data.id,s.loadTransactions(t,a)}))},loadTransactions:function(t,a){var s=this;axios.get("./api/v1/accounts/"+a+"/transactions?page=1&limit=10").then((function(a){s.accounts[t].transactions=a.data.data}))}}},h=Object(n.a)(v,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"row"},t._l(t.accounts,(function(a){return s("div",{class:{"col-lg-12":1===t.accounts.length,"col-lg-6":2===t.accounts.length,"col-lg-4":t.accounts.length>2}},[s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[s("a",{attrs:{href:a.uri}},[t._v(t._s(a.title))])])]),t._v(" "),s("div",{staticClass:"card-body table-responsive p-0"},[1===t.accounts.length?s("transaction-list-large",{attrs:{transactions:a.transactions,account_id:a.id}}):t._e(),t._v(" "),2===t.accounts.length?s("transaction-list-medium",{attrs:{transactions:a.transactions,account_id:a.id}}):t._e(),t._v(" "),t.accounts.length>2?s("transaction-list-small",{attrs:{transactions:a.transactions,account_id:a.id}}):t._e()],1)])])})),0)}),[],!1,null,"40cb1578",null).exports,p={name:"MainBillsList",mounted:function(){var t=this;axios.get("./api/v1/bills?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(a){t.loadBills(a.data.data)}))},components:{},methods:{loadBills:function(t){for(var a in t)if(t.hasOwnProperty(a)&&/^0$|^[1-9]\d*$/.test(a)&&a<=4294967294){var s=t[a],e=s.attributes.active;s.attributes.pay_dates.length>0&&e&&this.bills.push(s)}}},data:function(){return{bills:[]}},computed:{}},m=Object(n.a)(p,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.bills")))])]),t._v(" "),s("div",{staticClass:"card-body table-responsive p-0"},[s("table",{staticClass:"table table-striped"},[s("caption",{staticStyle:{display:"none"}},[t._v(t._s(t.$t("firefly.bills")))]),t._v(" "),s("thead",[s("tr",[s("th",{staticStyle:{width:"35%"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.name")))]),t._v(" "),s("th",{staticStyle:{width:"40%"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.amount")))]),t._v(" "),s("th",{staticStyle:{width:"25%"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.next_expected_match")))])])]),t._v(" "),s("tbody",t._l(this.bills,(function(a){return s("tr",[s("td",[s("a",{attrs:{href:"./bills/show"+a.id,title:a.attributes.name}},[t._v(t._s(a.attributes.name))])]),t._v(" "),s("td",[t._v("~"+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.attributes.currency_code}).format((a.attributes.amount_min+a.attributes.amount_max)/2))+"\n                ")]),t._v(" "),s("td",t._l(a.attributes.pay_dates,(function(a){return s("span",[t._v("\n                        "+t._s(a)),s("br")])})),0)])})),0)])]),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./bills"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_bills")))])])])}),[],!1,null,null,null).exports,b={name:"FormatLabel",methods:{formatLabel:function(t,a){var s=[],e=(t=String(t)).split(" "),n="";return e.forEach((function(t,i){if(n.length>0){var r=n+" "+t;if(!(r.length>a))return i===e.length-1?void s.push(r):void(n=r);s.push(n),n=""}i!==e.length-1&&t.length<a?n=t:s.push(t)})),s}}},y=Object(n.a)(b,void 0,void 0,!1,null,null,null).exports,g={name:"DefaultBarOptions",data:function(){return{}},methods:{getDefaultOptions:function(){return{type:"bar",layout:{padding:{left:50,right:50,top:0,bottom:0}},stacked:!0,elements:{line:{cubicInterpolationMode:"monotone"}},legend:{display:!1},animation:{duration:0},responsive:!0,maintainAspectRatio:!1,scales:{xAxes:[{stacked:!0,gridLines:{display:!1},ticks:{callback:function(t,a,s){return y.methods.formatLabel(t,20)}}}],yAxes:[{stacked:!1,display:!0,drawOnChartArea:!1,offset:!0,beginAtZero:!0,ticks:{callback:function(t){var a=this.chart.data.datasets[0]?this.chart.data.datasets[0].currency_code:"EUR";return new Intl.NumberFormat(window.localeValue,{style:"currency",currency:a}).format(t)}}}]},tooltips:{mode:"label",callbacks:{label:function(t,a){var s=a.datasets[t.datasetIndex]?a.datasets[t.datasetIndex].currency_code:"EUR",e=new Intl.NumberFormat(window.localeValue,{style:"currency",currency:s}).format(t.yLabel);return a.datasets[t.datasetIndex].label+": "+e}}}}}}},C=Object(n.a)(g,(function(){var t=this.$createElement;return(this._self._c||t)("div")}),[],!1,null,"ce29954a",null).exports,x={name:"MainBudget",mounted:function(){axios.get("./api/v1/chart/budget/overview?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(t){var a=l.methods.convertChart(t.data),s=$("#mainBudgetChart").get(0).getContext("2d");new Chart(s,{type:"bar",data:a,options:C.methods.getDefaultOptions()})}))}},w=Object(n.a)(x,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.budgets")))])]),t._v(" "),t._m(0),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./budgets"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_budgets")))])])])}),[function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"card-body"},[a("div",{staticStyle:{position:"relative"}},[a("canvas",{staticStyle:{"min-height":"400px",height:"400px","max-height":"400px","max-width":"100%"},attrs:{id:"mainBudgetChart"}})])])}],!1,null,"1092ec16",null).exports,S={name:"MainCategory",mounted:function(){axios.get("./api/v1/chart/category/overview?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(t){var a=l.methods.convertChart(t.data);a=l.methods.colorizeLineData(a);var s=$("#mainCategoryChart").get(0).getContext("2d");new Chart(s,{type:"bar",data:a,options:C.methods.getDefaultOptions()})}))}},E=Object(n.a)(S,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.categories")))])]),t._v(" "),t._m(0),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./categories"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_categories")))])])])}),[function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"card-body"},[a("div",[a("canvas",{staticStyle:{"min-height":"400px",height:"400px","max-height":"400px","max-width":"100%"},attrs:{id:"mainCategoryChart"}})])])}],!1,null,"0f053244",null).exports,D={name:"MainCredit",components:{},data:function(){return{transactions:[]}},mounted:function(){var t=this;axios.get("./api/v1/transactions?type=deposit&limit=10&start="+window.sessionStart+"&end="+window.sessionEnd).then((function(a){t.transactions=a.data.data}))},methods:{},computed:{}},k=Object(n.a)(D,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.income")))])]),t._v(" "),s("div",{staticClass:"card-body table-responsive p-0"},[s("transaction-list-small",{attrs:{transactions:this.transactions}})],1),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./accounts/revenue"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_deposits")))])])])}),[],!1,null,null,null).exports,O={name:"MainDebit",mounted:function(){axios.get("./api/v1/chart/account/expense?start="+window.sessionStart+"&end="+window.sessionEnd).then((function(t){var a=l.methods.convertChart(t.data);a=l.methods.colorizeLineData(a);var s=$("#mainDebitChart").get(0).getContext("2d");new Chart(s,{type:"bar",data:a,options:C.methods.getDefaultOptions()})}))}},I=Object(n.a)(O,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.expense_accounts")))])]),t._v(" "),t._m(0),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./transactions/withdrawals"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_expenses")))])])])}),[function(){var t=this.$createElement,a=this._self._c||t;return a("div",{staticClass:"card-body"},[a("div",[a("canvas",{staticStyle:{"min-height":"400px",height:"400px","max-height":"400px","max-width":"100%"},attrs:{id:"mainDebitChart"}})])])}],!1,null,"19053b36",null).exports,U={name:"MainPiggyList",mounted:function(){var t=this;axios.get("./api/v1/piggy_banks").then((function(a){t.loadPiggyBanks(a.data.data)}))},methods:{loadPiggyBanks:function(t){for(var a in t)if(t.hasOwnProperty(a)&&/^0$|^[1-9]\d*$/.test(a)&&a<=4294967294){var s=t[a];0!==parseFloat(s.attributes.left_to_save)&&(s.attributes.pct=parseFloat(s.attributes.current_amount)/parseFloat(s.attributes.target_amount)*100,this.piggy_banks.push(s))}this.piggy_banks.sort((function(t,a){return a.attributes.pct-t.attributes.pct}))}},data:function(){return{piggy_banks:[]}}},L=Object(n.a)(U,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("div",{staticClass:"card"},[s("div",{staticClass:"card-header"},[s("h3",{staticClass:"card-title"},[t._v(t._s(t.$t("firefly.piggy_banks")))])]),t._v(" "),s("div",{staticClass:"card-body table-responsive p-0"},[s("table",{staticClass:"table table-striped"},[s("caption",{staticStyle:{display:"none"}},[t._v(t._s(t.$t("firefly.piggy_banks")))]),t._v(" "),s("thead",[s("tr",[s("th",{staticStyle:{width:"35%"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.piggy_bank")))]),t._v(" "),s("th",{staticStyle:{width:"40%"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.percentage")))]),t._v(" "),s("th",{staticStyle:{width:"25%","text-align":"right"},attrs:{scope:"col"}},[t._v(t._s(t.$t("list.amount")))])])]),t._v(" "),s("tbody",t._l(this.piggy_banks,(function(a){return s("tr",[s("td",[t._v(t._s(a.attributes.name)+"\n                "),s("br"),s("small",{staticClass:"text-muted"},[t._v(t._s(a.attributes.object_group_title))])]),t._v(" "),s("td",[s("div",{staticClass:"progress-group"},[s("div",{staticClass:"progress progress-sm"},[a.attributes.pct<100?s("div",{staticClass:"progress-bar primary",style:{width:a.attributes.pct+"%"}}):t._e(),t._v(" "),100===a.attributes.pct?s("div",{staticClass:"progress-bar bg-success",style:{width:a.attributes.pct+"%"}}):t._e()])])]),t._v(" "),s("td",{staticStyle:{"text-align":"right"}},[s("span",{staticClass:"text-success"},[t._v("\n                        "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.attributes.currency_code}).format(a.attributes.current_amount))+"\n                    ")]),t._v("\n                    of\n                    "),s("span",{staticClass:"text-success"},[t._v(t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.attributes.currency_code}).format(a.attributes.target_amount)))])])])})),0)])]),t._v(" "),s("div",{staticClass:"card-footer"},[s("a",{staticClass:"btn btn-default button-sm",attrs:{href:"./piggy-banks"}},[s("i",{staticClass:"far fa-money-bill-alt"}),t._v(" "+t._s(t.$t("firefly.go_to_piggies")))])])])}),[],!1,null,"68217368",null).exports,A={name:"TransactionListLarge",props:{transactions:{type:Array,default:function(){return[]}},account_id:{type:Number,default:function(){return 0}}}},F=Object(n.a)(A,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("table",{staticClass:"table table-striped table-sm"},[s("caption",{staticStyle:{display:"none"}},[t._v(t._s(t.$t("firefly.transaction_table_description")))]),t._v(" "),s("thead",[s("tr",[s("th",{staticClass:"text-left",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.description")))]),t._v(" "),s("th",{attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.opposing_account")))]),t._v(" "),s("th",{staticClass:"text-right",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.amount")))]),t._v(" "),s("th",{attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.category")))]),t._v(" "),s("th",{attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.budget")))])])]),t._v(" "),s("tbody",t._l(this.transactions,(function(a){return s("tr",[s("td",[s("a",{attrs:{href:"transactions/show/"+a.id,title:a.date}},[a.attributes.transactions.length>1?s("span",[t._v(t._s(a.attributes.group_title))]):t._e(),t._v(" "),1===a.attributes.transactions.length?s("span",[t._v(t._s(a.attributes.transactions[0].description))]):t._e()])]),t._v(" "),s("td",t._l(a.attributes.transactions,(function(e){return s("span",["withdrawal"===e.type?s("a",{attrs:{href:"accounts/show/"+a.destination_id}},[t._v(t._s(e.destination_name))]):t._e(),t._v(" "),"deposit"===e.type?s("a",{attrs:{href:"accounts/show/"+a.source_id}},[t._v(t._s(e.source_name))]):t._e(),t._v(" "),"transfer"===e.type&&e.source_id===t.account_id?s("a",{attrs:{href:"accounts/show/"+a.destination_id}},[t._v(t._s(e.destination_name))]):t._e(),t._v(" "),"transfer"===e.type&&e.destination_id===t.account_id?s("a",{attrs:{href:"accounts/show/"+a.source_id}},[t._v(t._s(e.source_name))]):t._e(),t._v(" "),s("br")])})),0),t._v(" "),s("td",{staticStyle:{"text-align":"right"}},t._l(a.attributes.transactions,(function(a){return s("span",["withdrawal"===a.type?s("span",{staticClass:"text-danger"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"deposit"===a.type?s("span",{staticClass:"text-success"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.source_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.destination_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e()])})),0),t._v(" "),s("td",t._l(a.attributes.transactions,(function(e){return s("span",[0!==e.category_id?s("a",{attrs:{href:"categories/show/"+a.category_id}},[t._v(t._s(e.category_name))]):t._e(),s("br")])})),0),t._v(" "),s("td",t._l(a.attributes.transactions,(function(e){return s("span",[0!==e.budget_id?s("a",{attrs:{href:"budgets/show/"+a.budget_id}},[t._v(t._s(e.budget_name))]):t._e(),s("br")])})),0)])})),0)])}),[],!1,null,"6f6fc869",null).exports,N={name:"TransactionListMedium",props:{transactions:{type:Array,default:function(){return[]}},account_id:{type:Number,default:function(){return 0}}}},V=Object(n.a)(N,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("table",{staticClass:"table table-striped table-sm"},[s("caption",{staticStyle:{display:"none"}},[t._v(t._s(t.$t("firefly.transaction_table_description")))]),t._v(" "),s("thead",[s("tr",[s("th",{staticClass:"text-left",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.description")))]),t._v(" "),s("th",{attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.opposing_account")))]),t._v(" "),s("th",{staticClass:"text-right",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.amount")))])])]),t._v(" "),s("tbody",t._l(this.transactions,(function(a){return s("tr",[s("td",[s("a",{attrs:{href:"transactions/show/"+a.id,title:a.date}},[a.attributes.transactions.length>1?s("span",[t._v(t._s(a.attributes.group_title))]):t._e(),t._v(" "),1===a.attributes.transactions.length?s("span",[t._v(t._s(a.attributes.transactions[0].description))]):t._e()])]),t._v(" "),s("td",t._l(a.attributes.transactions,(function(e){return s("span",["withdrawal"===e.type?s("a",{attrs:{href:"accounts/show/"+a.destination_id}},[t._v(t._s(e.destination_name))]):t._e(),t._v(" "),"deposit"===e.type?s("a",{attrs:{href:"accounts/show/"+a.source_id}},[t._v(t._s(e.source_name))]):t._e(),t._v(" "),"transfer"===e.type&&e.source_id===t.account_id?s("a",{attrs:{href:"accounts/show/"+a.destination_id}},[t._v(t._s(e.destination_name))]):t._e(),t._v(" "),"transfer"===e.type&&e.destination_id===t.account_id?s("a",{attrs:{href:"accounts/show/"+a.source_id}},[t._v(t._s(e.source_name))]):t._e(),t._v(" "),s("br")])})),0),t._v(" "),s("td",{staticStyle:{"text-align":"right"}},t._l(a.attributes.transactions,(function(a){return s("span",["withdrawal"===a.type?s("span",{staticClass:"text-danger"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"deposit"===a.type?s("span",{staticClass:"text-success"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.source_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.destination_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e()])})),0)])})),0)])}),[],!1,null,"fd843822",null).exports,j={name:"TransactionListSmall",props:{transactions:{type:Array,default:function(){return[]}},account_id:{type:Number,default:function(){return 0}}}},P=Object(n.a)(j,(function(){var t=this,a=t.$createElement,s=t._self._c||a;return s("table",{staticClass:"table table-striped table-sm"},[s("caption",{staticStyle:{display:"none"}},[t._v(t._s(t.$t("firefly.transaction_table_description")))]),t._v(" "),s("thead",[s("tr",[s("th",{staticClass:"text-left",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.description")))]),t._v(" "),s("th",{staticClass:"text-right",attrs:{scope:"col"}},[t._v(t._s(t.$t("firefly.amount")))])])]),t._v(" "),s("tbody",t._l(this.transactions,(function(a){return s("tr",[s("td",[s("a",{attrs:{href:"transactions/show/"+a.id,title:a.date}},[a.attributes.transactions.length>1?s("span",[t._v(t._s(a.attributes.group_title))]):t._e(),t._v(" "),1===a.attributes.transactions.length?s("span",[t._v(t._s(a.attributes.transactions[0].description))]):t._e()])]),t._v(" "),s("td",{staticStyle:{"text-align":"right"}},t._l(a.attributes.transactions,(function(a){return s("span",["withdrawal"===a.type?s("span",{staticClass:"text-danger"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"deposit"===a.type?s("span",{staticClass:"text-success"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.source_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(-1*a.amount))),s("br")]):t._e(),t._v(" "),"transfer"===a.type&&a.destination_id===t.account_id?s("span",{staticClass:"text-info"},[t._v("\n                    "+t._s(Intl.NumberFormat("en-US",{style:"currency",currency:a.currency_code}).format(a.amount))),s("br")]):t._e()])})),0)])})),0)])}),[],!1,null,"2e914636",null).exports;s(20),Vue.component("transaction-list-large",F),Vue.component("transaction-list-medium",V),Vue.component("transaction-list-small",P),Vue.component("dashboard",i),Vue.component("top-boxes",c),Vue.component("main-account",f),Vue.component("main-account-list",h),Vue.component("main-bills-list",m),Vue.component("main-budget",w),Vue.component("main-category",E),Vue.component("main-credit",k),Vue.component("main-debit",I),Vue.component("main-piggy-list",L);var T=s(18),B={};new Vue({i18n:T,el:"#dashboard",render:function(t){return t(i,{props:B})}})},74:function(t,a){}},[[21,0,1]]]);
+(window.webpackJsonp=window.webpackJsonp||[]).push([[3],{169:function(s,j,n){n(170),s.exports=n(226)},170:function(s,j,n){"use strict";n.r(j);var e=n(8),t=n(12),a=n(11),o=n(13),r=n(14),i=n(15),m=n(16),c=n(17),u=n(18),l=n(19),d=n(20),h=n(21),p=n(22);n(10),n(195),Vue.component("transaction-list-large",d.a),Vue.component("transaction-list-medium",h.a),Vue.component("transaction-list-small",p.a),Vue.component("dashboard",e.a),Vue.component("top-boxes",t.a),Vue.component("main-account",a.a),Vue.component("main-account-list",o.a),Vue.component("main-bills-list",r.a),Vue.component("main-budget",i.a),Vue.component("main-category",m.a),Vue.component("main-credit",c.a),Vue.component("main-debit",u.a),Vue.component("main-piggy-list",l.a);var b=n(167),k={};new Vue({i18n:b,el:"#dashboard",render:function(s){return s(e.a,{props:k})}})},197:function(s,j,n){var e={"./af":34,"./af.js":34,"./ar":35,"./ar-dz":36,"./ar-dz.js":36,"./ar-kw":37,"./ar-kw.js":37,"./ar-ly":38,"./ar-ly.js":38,"./ar-ma":39,"./ar-ma.js":39,"./ar-sa":40,"./ar-sa.js":40,"./ar-tn":41,"./ar-tn.js":41,"./ar.js":35,"./az":42,"./az.js":42,"./be":43,"./be.js":43,"./bg":44,"./bg.js":44,"./bm":45,"./bm.js":45,"./bn":46,"./bn.js":46,"./bo":47,"./bo.js":47,"./br":48,"./br.js":48,"./bs":49,"./bs.js":49,"./ca":50,"./ca.js":50,"./cs":51,"./cs.js":51,"./cv":52,"./cv.js":52,"./cy":53,"./cy.js":53,"./da":54,"./da.js":54,"./de":55,"./de-at":56,"./de-at.js":56,"./de-ch":57,"./de-ch.js":57,"./de.js":55,"./dv":58,"./dv.js":58,"./el":59,"./el.js":59,"./en-au":60,"./en-au.js":60,"./en-ca":61,"./en-ca.js":61,"./en-gb":62,"./en-gb.js":62,"./en-ie":63,"./en-ie.js":63,"./en-il":64,"./en-il.js":64,"./en-in":65,"./en-in.js":65,"./en-nz":66,"./en-nz.js":66,"./en-sg":67,"./en-sg.js":67,"./eo":68,"./eo.js":68,"./es":69,"./es-do":70,"./es-do.js":70,"./es-us":71,"./es-us.js":71,"./es.js":69,"./et":72,"./et.js":72,"./eu":73,"./eu.js":73,"./fa":74,"./fa.js":74,"./fi":75,"./fi.js":75,"./fil":76,"./fil.js":76,"./fo":77,"./fo.js":77,"./fr":78,"./fr-ca":79,"./fr-ca.js":79,"./fr-ch":80,"./fr-ch.js":80,"./fr.js":78,"./fy":81,"./fy.js":81,"./ga":82,"./ga.js":82,"./gd":83,"./gd.js":83,"./gl":84,"./gl.js":84,"./gom-deva":85,"./gom-deva.js":85,"./gom-latn":86,"./gom-latn.js":86,"./gu":87,"./gu.js":87,"./he":88,"./he.js":88,"./hi":89,"./hi.js":89,"./hr":90,"./hr.js":90,"./hu":91,"./hu.js":91,"./hy-am":92,"./hy-am.js":92,"./id":93,"./id.js":93,"./is":94,"./is.js":94,"./it":95,"./it-ch":96,"./it-ch.js":96,"./it.js":95,"./ja":97,"./ja.js":97,"./jv":98,"./jv.js":98,"./ka":99,"./ka.js":99,"./kk":100,"./kk.js":100,"./km":101,"./km.js":101,"./kn":102,"./kn.js":102,"./ko":103,"./ko.js":103,"./ku":104,"./ku.js":104,"./ky":105,"./ky.js":105,"./lb":106,"./lb.js":106,"./lo":107,"./lo.js":107,"./lt":108,"./lt.js":108,"./lv":109,"./lv.js":109,"./me":110,"./me.js":110,"./mi":111,"./mi.js":111,"./mk":112,"./mk.js":112,"./ml":113,"./ml.js":113,"./mn":114,"./mn.js":114,"./mr":115,"./mr.js":115,"./ms":116,"./ms-my":117,"./ms-my.js":117,"./ms.js":116,"./mt":118,"./mt.js":118,"./my":119,"./my.js":119,"./nb":120,"./nb.js":120,"./ne":121,"./ne.js":121,"./nl":122,"./nl-be":123,"./nl-be.js":123,"./nl.js":122,"./nn":124,"./nn.js":124,"./oc-lnc":125,"./oc-lnc.js":125,"./pa-in":126,"./pa-in.js":126,"./pl":127,"./pl.js":127,"./pt":128,"./pt-br":129,"./pt-br.js":129,"./pt.js":128,"./ro":130,"./ro.js":130,"./ru":131,"./ru.js":131,"./sd":132,"./sd.js":132,"./se":133,"./se.js":133,"./si":134,"./si.js":134,"./sk":135,"./sk.js":135,"./sl":136,"./sl.js":136,"./sq":137,"./sq.js":137,"./sr":138,"./sr-cyrl":139,"./sr-cyrl.js":139,"./sr.js":138,"./ss":140,"./ss.js":140,"./sv":141,"./sv.js":141,"./sw":142,"./sw.js":142,"./ta":143,"./ta.js":143,"./te":144,"./te.js":144,"./tet":145,"./tet.js":145,"./tg":146,"./tg.js":146,"./th":147,"./th.js":147,"./tk":148,"./tk.js":148,"./tl-ph":149,"./tl-ph.js":149,"./tlh":150,"./tlh.js":150,"./tr":151,"./tr.js":151,"./tzl":152,"./tzl.js":152,"./tzm":153,"./tzm-latn":154,"./tzm-latn.js":154,"./tzm.js":153,"./ug-cn":155,"./ug-cn.js":155,"./uk":156,"./uk.js":156,"./ur":157,"./ur.js":157,"./uz":158,"./uz-latn":159,"./uz-latn.js":159,"./uz.js":158,"./vi":160,"./vi.js":160,"./x-pseudo":161,"./x-pseudo.js":161,"./yo":162,"./yo.js":162,"./zh-cn":163,"./zh-cn.js":163,"./zh-hk":164,"./zh-hk.js":164,"./zh-mo":165,"./zh-mo.js":165,"./zh-tw":166,"./zh-tw.js":166};function t(s){var j=a(s);return n(j)}function a(s){if(!n.o(e,s)){var j=new Error("Cannot find module '"+s+"'");throw j.code="MODULE_NOT_FOUND",j}return e[s]}t.keys=function(){return Object.keys(e)},t.resolve=a,s.exports=t,t.id=197},226:function(s,j){}},[[169,0,1]]]);
 //# sourceMappingURL=dashboard.js.map
