@@ -23,5 +23,5 @@
 window.$ = window.jQuery = require('jquery');
 
 // admin stuff
-require('bootstrap');
+//require('bootstrap');
 
