@@ -233,6 +233,7 @@ return [
     'all_destination_accounts'                            => 'Cuentas destino',
     'all_source_accounts'                                 => 'Cuentas origen',
     'back_to_index'                                       => 'Volver al Índice',
+    'cant_logout_guard'                                   => 'Firefly III can\'t log you out.',
 
     // check for updates:
     'update_check_title'                                  => 'Ver actualizaciones',
@@ -1008,6 +1009,7 @@ return [
     'all_accounts_inactive'                     => 'Estas son sus cuentas inactivas.',
     'active_account_link'                       => 'Este enlace se regresa a sus cuentas activas.',
     'account_missing_transaction'               => 'Cuenta #:id (":name") no puede ser vista directamente, Firefly perdió información de redirección.',
+    'cc_monthly_payment_date_help'              => 'Select any year and any month, it will be ignored anyway. Only the day of the month is relevant.',
     'details_for_asset'                         => 'Detalles para la cuenta de activos ":name"',
     'details_for_expense'                       => 'Detalles para la cuenta de gastos ":name"',
     'details_for_revenue'                       => 'Detalles para la cuenta de ingresos ":name"',
@@ -1629,7 +1631,7 @@ return [
     'after_update_create_another'           => 'Después de actualizar, vuelve aquí para continuar editando.',
     'store_as_new'                          => 'Almacenar como una nueva transacción en lugar de actualizar.',
     'reset_after'                           => 'Restablecer formulario después del envío',
-    'errors_submission'                     => 'Hubo algo malo con su envío. Por favor, revise los errores de abajo.',
+    'errors_submission'                     => 'There was something wrong with your submission. Please check out the errors.',
 
     // object groups
     'default_group_title_name'              => '(sin agrupación)',
