@@ -24,20 +24,20 @@ declare(strict_types=1);
 
 return [
     'html_language'           => 'fr',
-    'locale'                  => 'fr_FR.utf8, fr_FR.UTF-8',
+    'locale'                  => 'fr, French, fr_FR.utf8, fr_FR.UTF-8',
     'month'                   => '%B %Y',
-    'month_and_day'           => '%e %B %Y',
+    'month_and_day'           => '%B %e %Y',
     'month_and_day_moment_js' => 'D MMMM YYYY',
     'month_and_date_day'      => '%A %e %B %Y',
     'month_and_day_no_year'   => '%e %B',
-    'date_time'               => '%B %e %Y @ %T',
+    'date_time'               => '%e %B %Y à %T',
     'specific_day'            => '%e %B %Y',
     'week_in_year'            => 'Semaine %V, %G',
     'year'                    => '%Y',
     'half_year'               => '%B %Y',
     'month_js'                => 'MMMM YYYY',
     'month_and_day_js'        => 'Do MMMM YYYY',
-    'date_time_js'            => 'Do MMMM YYYY, @ HH:mm:ss',
+    'date_time_js'            => 'Do MMMM YYYY, à HH:mm:ss',
     'specific_day_js'         => 'D MMMM YYYY',
     'week_in_year_js'         => '[Week] w, YYYY',
     'year_js'                 => 'YYYY',
