@@ -233,6 +233,7 @@ return [
     'all_destination_accounts'                            => 'Cieľové účty',
     'all_source_accounts'                                 => 'Zdrojové účty',
     'back_to_index'                                       => 'Späť na zoznam',
+    'cant_logout_guard'                                   => 'Firefly III can\'t log you out.',
 
     // check for updates:
     'update_check_title'                                  => 'Zistiť dostupnosť aktualizácií',
@@ -1008,6 +1009,7 @@ return [
     'all_accounts_inactive'                     => 'Toto sú vaše neaktívne bankové účty.',
     'active_account_link'                       => 'Tento prepojenie smeruje späť na vaše aktívne účty.',
     'account_missing_transaction'               => 'Bankový účet #:id (":name") nie je možné zobraziť priamo, ale Firefly III chýbajú informácie o presmerovaní.',
+    'cc_monthly_payment_date_help'              => 'Select any year and any month, it will be ignored anyway. Only the day of the month is relevant.',
     'details_for_asset'                         => 'Podrobnosti o účte aktív „:name“',
     'details_for_expense'                       => 'Podrobnosti o výdavkovom účte „:name“',
     'details_for_revenue'                       => 'Podrobnosti o výnosovom účte „:name“',
@@ -1629,7 +1631,7 @@ return [
     'after_update_create_another'           => 'Po aktualizácii sa vrátiť späť a pokračovať v úpravách.',
     'store_as_new'                          => 'Namiesto aktualizácie uložiť ako novú transakciu.',
     'reset_after'                           => 'Po odoslaní vynulovať formulár',
-    'errors_submission'                     => 'Pri odoslaní nastala chyba. Skontrolujte chyby uvedené nižšie.',
+    'errors_submission'                     => 'There was something wrong with your submission. Please check out the errors.',
 
     // object groups
     'default_group_title_name'              => '(nezoskupené)',
